1. Creating our project
Install vue-cli:
sudo npm update -g @vue/cli

Create project:
vue create real-world-vue

Run development server:
yarn run serve

Repeat the same process as when i created the project with vue ui GUI by running:
vue ui

Vue UI Features
The UI provides the ability to alter our configurations, as well as run tasks, like the serve task that we ran earlier from the terminal. The difference here is that we get a lot more information about what is happening with that task.
You can also add plugins to your project from the Vue UI, which makes it very simple to add a library that you may need.
If you need to import a project that you hadn’t originally created from the Vue UI, you can easily do so from the Import tab of the Project Manager. Just locate your project, and click Import this folder.

Touring our Vue Project
Now that we know how to create our project from the terminal and also from the UI, let’s take a look at the project that was created for us.
The node_modules directory is where all of the libraries we need to build Vue are stored.
In the public directory, you place any static assets you don’t want to be run through Webpack when we build our project.
You’ll want to put the majority of your assets, such as images and fonts, in the assets directory so they can be optimized by Webpack.
The components directory is where we store the components, or building blocks, of our Vue app.
The views directory is where we store files for the different views, or pages, of our app.
The App.vue file is the root component that all other components are nested within.
The main.js file is what renders our App.vue component (and everything nested within it) and mounts it to the DOM.
Below that we have a router.js file, which we’ll cover more in the lesson on Vue Router, and the store.js file is for Vuex, which we’ll also cover in future lessons.
Finally, we have a .gitignore file where we can specify what we want git to ignore, along with a babel.config.js file and our package.json, which helps npm identify the project and handle its dependencies.