3. Optimizing your editor (Installing ESLint & Prettier; Configuring ESLint; Configuring Prettier) Installing ESLint & Prettier Now, we need to make sure we have ESLint and Prettier installed. In the extensions store, we’ll do a search for ESLint, then go ahead and install it. And we’ll do the same for Prettier. Once it’s installed, we’ll hit reload to reload VS Code.
Configuring ESLint
Now that these are installed, we need to add a bit of extra configuration to them.
When we created our project from the terminal, we chose to create it with dedicated config files, which gave us this .eslintrc.js file, where we can configure ESLint for this project. Had we not chosen dedicated files, we would find the ESLint configurations within our package.json.
So, in our .eslintrc.js file, we’ll add:
'plugin:prettier/recommended'
This will enable Prettier support in ESLint with the default settings.

Configuring Prettier
We also have the option to create a Prettier configuration file, to add some special settings according to our personal style or our team’s preferences.
We’ll create it here and name it .prettierrc.js.
This will convert double quotes to single quotes, and make sure that semicolons are not automatically inserted.